Renovate implementation of MCExtractor::writeInfo
The method has been basically rewritten while keeping its functionality
(mostly).

Selection of particles to be stored is updated to use status flags of
reco::GenParticle. Now the extractor is supposed to work correctly with
Pythia 8, and this fixes #1. The list of particles stored has been
extended a bit, adding prompt electrons and muons as well as leptons
from decay of a J/psi.

Information about mothers is now stored in a better way: indices of
ancestor particles correspond directly to positions in the arrays in the
output tree (and not to indices in the original collection in AOD).

A number of stored arrays is redundant. For the time being, they are
kept for backward compatibility. However, plugins are only expected to
access information on generator-level particles through the interface of
MCExtractor, these redundant branches can be removed already now.