Store also quark decay products of certain resonances in MCExtractor
This is important to catch hadronic decays of W boson. Here is a list of
light-flavour quarks in a semileptonic ttbar events (MadGraph with
Pythia 8, MiniAOD):

 PDG ID: 1, status: 23, mother PDG ID: -24, fromHardProcessBeforeFSR? 0
 PDG ID: -4, status: 23, mother PDG ID: -24, fromHardProcessBeforeFSR? 1
 PDG ID: 1, status: 71, mother PDG ID: 1, fromHardProcessBeforeFSR? 0
 PDG ID: -2, status: 71, mother PDG ID: -4, fromHardProcessBeforeFSR? 0
 PDG ID: 4, status: 71, mother PDG ID: 2212, fromHardProcessBeforeFSR? 0
 PDG ID: 2, status: 71, mother PDG ID: -4, fromHardProcessBeforeFSR? 0
 PDG ID: -4, status: 71, mother PDG ID: -4, fromHardProcessBeforeFSR? 0
 PDG ID: 2, status: 71, mother PDG ID: 2212, fromHardProcessBeforeFSR? 0
 PDG ID: -4, status: 71, mother PDG ID: 2212, fromHardProcessBeforeFSR? 0
 PDG ID: -1, status: 71, mother PDG ID: 2212, fromHardProcessBeforeFSR? 0
 PDG ID: -3, status: 71, mother PDG ID: 2212, fromHardProcessBeforeFSR? 0
 PDG ID: -3, status: 71, mother PDG ID: 2212, fromHardProcessBeforeFSR? 0

Usually, both quarks from the decay of W boson appear in the history
with the fromHardProcessBeforeFSR flag set to true, but in the above
event the up quark, and all its subsequent duplicates stored in MiniAOD,
have the flag set to false. Without this fix, such quarks would not be
saved.